refactor(collection): remove references to "workbench" in favor of "collection"
Rename previous `Workbench.tsx` file to `Collection.tsx`. Rename previous `Collection.tsx` to `CollectionHome.tsx`.

rename `workbench` directory to `colleciton`, and move the contents of the previous `colleciton` into the new `collection` directory.

With more eyes on this, wanted to take the opportunity to move components around in a way that would be easier to understand for folks who don't know historically how the app was shaped. Now, most components called in the Collection view is now in the `collection` directory.

In future refactors we can move dataset-only functionality to lower components, and hopefully streamline logic.
