+++ app/components/collection/ComponentRouter.tsx
@@ -19,19 +19,19 @@ import { setRecentEditRef, setRecentDatasetRef } from '../../actions/workbenchRo
 
 import { selectWorkingDatasetIsLoading, selectWorkingDataset, selectFsiPath, selectDatasetIsLoading, selectDatasetStatusInfo, selectStatusInfoFromMutations } from '../../selections'
 
-import Body from './components/Body'
+import Body from './datasetComponents/Body'
 import CalloutBlock from '../chrome/CalloutBlock'
-import CommitEditor from './components/CommitEditor'
-import Commit from './components/Commit'
+import CommitEditor from './datasetComponents/CommitEditor'
+import Commit from './datasetComponents/Commit'
 import DropZone from '../chrome/DropZone'
 import Icon from '../chrome/Icon'
-import Metadata from './components/Metadata'
-import MetadataEditor from './components/MetadataEditor'
-import ReadmeEditor from './components/ReadmeEditor'
-import Readme from './components/Readme'
+import Metadata from './datasetComponents/Metadata'
+import MetadataEditor from './datasetComponents/MetadataEditor'
+import ReadmeEditor from './datasetComponents/ReadmeEditor'
+import Readme from './datasetComponents/Readme'
 import Structure from '../Structure'
-import Transform from './components/Transform'
-import StructureEditor from './components/StructureEditor'
+import Transform from './datasetComponents/Transform'
+import StructureEditor from './datasetComponents/StructureEditor'
 import StatusDot from '../chrome/StatusDot'
 import { getComponentDisplayProps } from './WorkingComponentList'
