+++ app/components/collection/collectionHome/DatasetList.tsx
@@ -5,19 +5,19 @@ import { shell, MenuItemConstructorOptions } from 'electron'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faTimes } from '@fortawesome/free-solid-svg-icons'
 
-import { QriRef } from '../../models/qriRef'
-import { Modal, ModalType } from '../../models/modals'
-import { MyDatasets, WorkingDataset, VersionInfo, RouteProps } from '../../models/store'
-import { connectComponentToPropsWithRouter } from '../../utils/connectComponentToProps'
-import { selectImportFileName, selectImportFileSize } from '../../selections'
-import { setFilter } from '../../actions/myDatasets'
-import { fetchMyDatasets } from '../../actions/api'
-import { setWorkingDataset } from '../../actions/selections'
-import { setModal } from '../../actions/ui'
+import { QriRef } from '../../../models/qriRef'
+import { Modal, ModalType } from '../../../models/modals'
+import { MyDatasets, WorkingDataset, VersionInfo, RouteProps } from '../../../models/store'
+import { connectComponentToPropsWithRouter } from '../../../utils/connectComponentToProps'
+import { selectImportFileName, selectImportFileSize } from '../../../selections'
+import { setFilter } from '../../../actions/myDatasets'
+import { fetchMyDatasets } from '../../../actions/api'
+import { setWorkingDataset } from '../../../actions/selections'
+import { setModal } from '../../../actions/ui'
 
-import ProgressBar from '../chrome/ProgressBar'
-import VersionInfoItem from '../item/VersionInfoItem'
-import { pathToDataset } from '../../paths'
+import ProgressBar from '../../chrome/ProgressBar'
+import VersionInfoItem from '../../item/VersionInfoItem'
+import { pathToDataset } from '../../../paths'
 
 // for displaying a progress bar based on import file size
 // assumes an import rate of 4828 bytes per millisecond
