+++ app/components/collection/datasetComponents/StructureEditor.tsx
